change foder location of comps
